Split repo into two parts for usability testing
